+++ lib/whois/record/scanners/whois.domainregistry.ie.rb
@@ -20,6 +20,7 @@ module Whois
         self.tokenizers += [
             :skip_empty_line,
             :scan_copyright,
+            :scan_contact,
             :scan_keyvalue,
             :scan_available,
         ]
@@ -36,7 +37,17 @@ module Whois
             while @input.scan(/^%(.+)\n/)
               lines << @input[1].strip
             end
-            @ast["field:disclaimer"] = lines.join(" ")
+            @ast["field:disclaimer"] = lines.join("\n")
+          end
+        end
+
+        tokenizer :scan_contact do
+          if @input.match?(/^person:/)
+            lines = {}
+            while @input.scan(/(.+?):(.*?)\n/)
+              lines.merge! @input[1].strip => @input[2].strip
+            end
+            @ast["field:#{lines['nic-hdl']}"] = lines
           end
         end